+++ tests/src/rules/no-anonymous-default-export.js
@@ -1,4 +1,4 @@
-import { test, SYNTAX_CASES } from '../utils';
+import { test, testVersion, SYNTAX_CASES } from '../utils';
 
 import { RuleTester } from 'eslint';
 
@@ -6,7 +6,7 @@ const ruleTester = new RuleTester();
 const rule = require('rules/no-anonymous-default-export');
 
 ruleTester.run('no-anonymous-default-export', rule, {
-  valid: [
+  valid: [].concat(
     // Exports with identifiers are valid
     test({ code: 'const foo = 123\nexport default foo' }),
     test({ code: 'export default function foo() {}' }),
@@ -31,12 +31,17 @@ ruleTester.run('no-anonymous-default-export', rule, {
     test({ code: 'export * from \'foo\'' }),
     test({ code: 'const foo = 123\nexport { foo }' }),
     test({ code: 'const foo = 123\nexport { foo as default }' }),
+    // es2022: Arbitrary module namespace identifier names
+    testVersion('>= 8.7', () => ({
+      code: 'const foo = 123\nexport { foo as "default" }',
+      parserOptions: { ecmaVersion: 2022 },
+    })),
 
     // Allow call expressions by default for backwards compatibility
     test({ code: 'export default foo(bar)' }),
 
     ...SYNTAX_CASES,
-  ],
+  ),
 
   invalid: [
     test({ code: 'export default []', errors: [{ message: 'Assign array to a variable before exporting as module default' }] }),