+++ tests/src/rules/no-named-as-default.js
@@ -1,11 +1,11 @@
-import { test, SYNTAX_CASES } from '../utils';
+import { test, testVersion, SYNTAX_CASES } from '../utils';
 import { RuleTester } from 'eslint';
 
 const ruleTester = new RuleTester();
 const rule = require('rules/no-named-as-default');
 
 ruleTester.run('no-named-as-default', rule, {
-  valid: [
+  valid: [].concat(
     test({ code: 'import "./malformed.js"' }),
 
     test({ code: 'import bar, { foo } from "./bar";' }),
@@ -21,10 +21,16 @@ ruleTester.run('no-named-as-default', rule, {
     test({ code: 'export default from "./bar";',
       parser: require.resolve('babel-eslint') }),
 
-    ...SYNTAX_CASES,
-  ],
+    // es2022: Arbitrary module namespae identifier names
+    testVersion('>= 8.7', () => ({
+      code: 'import bar, { foo } from "./export-default-string-and-named"',
+      parserOptions: { ecmaVersion: 2022 },
+    })),
 
-  invalid: [
+    ...SYNTAX_CASES,
+  ),
+
+  invalid: [].concat(
     test({
       code: 'import foo from "./bar";',
       errors: [ {
@@ -57,5 +63,23 @@ ruleTester.run('no-named-as-default', rule, {
         type: 'Literal',
       }],
     }),
-  ],
+
+    // es2022: Arbitrary module namespae identifier names
+    testVersion('>= 8.7', () => ({
+      code: 'import foo from "./export-default-string-and-named"',
+      errors: [{
+        message: 'Using exported name \'foo\' as identifier for default export.',
+        type: 'ImportDefaultSpecifier',
+      }],
+      parserOptions: { ecmaVersion: 2022 },
+    })),
+    testVersion('>= 8.7', () => ({
+      code: 'import foo, { foo as bar } from "./export-default-string-and-named"',
+      errors: [{
+        message: 'Using exported name \'foo\' as identifier for default export.',
+        type: 'ImportDefaultSpecifier',
+      }],
+      parserOptions: { ecmaVersion: 2022 },
+    })),
+  ),
 });